implement Dockerfile and refactor aced_install

+++ aced_install.sh
@@ -1,101 +1,11 @@
 #!/bin/bash
 
 TMP_FOLDER=$(mktemp -d)
-CONFIG_FILE='aced.conf'
-CONFIGFOLDER='/root/.acedcore'
-COIN_DAEMON='/usr/local/bin/acedd'
-COIN_CLI='/usr/local/bin/aced-cli'
-COIN_REPO='https://github.com/Acedcoin/AceD/releases/download/1.5/ubuntu16mn.tar.gz'
-#SENTINEL_REPO='https://github.com/cryptosharks131/sentinel'
-COIN_NAME='AceD'
-COIN_PORT=24126
-COIN_BS='https://github.com/Acedcoin/AceD/releases/download/1.5/acedCore.zip'
-
-
 NODEIP=$(curl -s4 icanhazip.com)
 
-
-RED='\033[0;31m'
-GREEN='\033[0;32m'
-NC='\033[0m'
-
-
-function install_sentinel() {
-  echo -e "${GREEN}Install sentinel.${NC}"
-  apt-get -y install python-virtualenv virtualenv >/dev/null 2>&1
-  git clone $SENTINEL_REPO /sentinel >/dev/null 2>&1
-  cd /sentinel
-  virtualenv ./venv >/dev/null 2>&1
-  ./venv/bin/pip install -r requirements.txt >/dev/null 2>&1
-  echo  "* * * * * cd /sentinel && ./venv/bin/python bin/sentinel.py >> $CONFIGFOLDER/sentinel.log 2>&1" > $CONFIGFOLDER/$COIN_NAME.cron
-  crontab $CONFIGFOLDER/$COIN_NAME.cron
-  rm $CONFIGFOLDER/$COIN_NAME.cron >/dev/null 2>&1
-  cd -
-}
-
-
-function compile_node() {
-  echo -e "Prepare to download $COIN_NAME"
-  cd $TMP_FOLDER
-  wget -q $COIN_REPO
-  compile_error
-  COIN_ZIP=$(echo $COIN_REPO | awk -F'/' '{print $NF}')
-  tar xvf $COIN_ZIP --strip 1 >/dev/null 2>&1
-  compile_error
-  cp aced{d,-cli} /usr/local/bin
-  compile_error
-  strip $COIN_DAEMON $COIN_CLI
-  cd - >/dev/null 2>&1
-  rm -rf $TMP_FOLDER >/dev/null 2>&1
-  chmod +x /usr/local/bin/acedd
-  chmod +x /usr/local/bin/aced-cli
-  clear
-}
-
-function configure_systemd() {
-  cat << EOF > /etc/systemd/system/$COIN_NAME.service
-[Unit]
-Description=$COIN_NAME service
-After=network.target
-
-[Service]
-User=root
-Group=root
-
-Type=forking
-#PIDFile=$CONFIGFOLDER/$COIN_NAME.pid
-
-ExecStart=$COIN_DAEMON -daemon -conf=$CONFIGFOLDER/$CONFIG_FILE -datadir=$CONFIGFOLDER
-ExecStop=-$COIN_CLI -conf=$CONFIGFOLDER/$CONFIG_FILE -datadir=$CONFIGFOLDER stop
-
-Restart=always
-PrivateTmp=true
-TimeoutStopSec=60s
-TimeoutStartSec=10s
-StartLimitInterval=120s
-StartLimitBurst=5
-
-[Install]
-WantedBy=multi-user.target
-EOF
-
-  systemctl daemon-reload
-  sleep 3
-  systemctl start $COIN_NAME.service
-  systemctl enable $COIN_NAME.service >/dev/null 2>&1
-
-  if [[ -z "$(ps axo cmd:100 | egrep $COIN_DAEMON)" ]]; then
-    echo -e "${RED}$COIN_NAME is not running${NC}, please investigate. You should start by running the following commands as root:"
-    echo -e "${GREEN}systemctl start $COIN_NAME.service"
-    echo -e "systemctl status $COIN_NAME.service"
-    echo -e "less /var/log/syslog${NC}"
-    exit 1
-  fi
-}
-
-
 function create_config() {
-  mkdir $CONFIGFOLDER >/dev/null 2>&1
+  echo -e "Creating Config..."
+  mkdir $CONFIGFOLDER
   RPCUSER=$(tr -cd '[:alnum:]' < /dev/urandom | fold -w10 | head -n1)
   RPCPASSWORD=$(tr -cd '[:alnum:]' < /dev/urandom | fold -w22 | head -n1)
   cat << EOF > $CONFIGFOLDER/$CONFIG_FILE
@@ -109,29 +19,8 @@ port=$COIN_PORT
 EOF
 }
 
-function create_key() {
-  echo -e "Enter your ${RED}$COIN_NAME Masternode Private Key${NC}. Leave it blank to generate a new ${RED}Masternode Private Key${NC} for you:"
-  read -e COINKEY
-  if [[ -z "$COINKEY" ]]; then
-  $COIN_DAEMON -daemon
-  sleep 30
-  if [ -z "$(ps axo cmd:100 | grep $COIN_DAEMON)" ]; then
-   echo -e "${RED}$COIN_NAME server couldn not start. Check /var/log/syslog for errors.{$NC}"
-   exit 1
-  fi
-  COINKEY=$($COIN_CLI masternode genkey)
-  if [ "$?" -gt "0" ];
-    then
-    echo -e "${RED}Wallet not fully loaded. Let us wait and try again to generate the Private Key${NC}"
-    sleep 30
-    COINKEY=$($COIN_CLI masternode genkey)
-  fi
-  $COIN_CLI stop
-fi
-clear
-}
-
 function update_config() {
+  echo -e "Updating Config..."
   sed -i 's/daemon=1/daemon=0/' $CONFIGFOLDER/$CONFIG_FILE
   cat << EOF >> $CONFIGFOLDER/$CONFIG_FILE
 logintimestamps=1
@@ -142,18 +31,6 @@ externalip=$NODEIP:$COIN_PORT
 masternodeprivkey=$COINKEY
 EOF
 }
-
-
-function enable_firewall() {
-  echo -e "Installing and setting up firewall to allow ingress on port ${GREEN}$COIN_PORT${NC}"
-  ufw allow $COIN_PORT/tcp comment "$COIN_NAME MN port" >/dev/null
-  ufw allow ssh comment "SSH" >/dev/null 2>&1
-  ufw limit ssh/tcp >/dev/null 2>&1
-  ufw default allow outgoing >/dev/null 2>&1
-  echo "y" | ufw enable >/dev/null 2>&1
-}
-
-
 
 function get_ip() {
   declare -a NODE_IPS
@@ -175,9 +52,9 @@ function get_ip() {
       NODEIP=${NODE_IPS[$choose_ip]}
   else
     NODEIP=${NODE_IPS[0]}
+    echo "IP Address.. $NODEIP"
   fi
 }
-
 
 function compile_error() {
 if [ "$?" -gt "0" ];
@@ -186,56 +63,6 @@ if [ "$?" -gt "0" ];
   exit 1
 fi
 }
-
-
-function checks() {
-if [[ $(lsb_release -d) != *16.04* ]]; then
-  echo -e "${RED}You are not running Ubuntu 16.04. Installation is cancelled.${NC}"
-  exit 1
-fi
-
-if [[ $EUID -ne 0 ]]; then
-   echo -e "${RED}$0 must be run as root.${NC}"
-   exit 1
-fi
-
-if [ -n "$(pidof $COIN_DAEMON)" ] || [ -e "$COIN_DAEMOM" ] ; then
-  echo -e "${RED}$COIN_NAME is already installed.${NC}"
-  exit 1
-fi
-}
-
-function prepare_system() {
-echo -e "Preparing the system to install ${GREEN}$COIN_NAME${NC} master node."
-echo -e "This might take 15-20 minutes and the screen will not move, so please be patient."
-apt-get update >/dev/null 2>&1
-DEBIAN_FRONTEND=noninteractive apt-get update > /dev/null 2>&1
-DEBIAN_FRONTEND=noninteractive apt-get -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" -y -qq upgrade >/dev/null 2>&1
-apt install -y software-properties-common >/dev/null 2>&1
-echo -e "${GREEN}Adding bitcoin PPA repository"
-apt-add-repository -y ppa:bitcoin/bitcoin >/dev/null 2>&1
-echo -e "Installing required packages, it may take some time to finish.${NC}"
-apt-get update >/dev/null 2>&1
-apt-get install -y -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" make software-properties-common \
-build-essential libtool autoconf libssl-dev libboost-dev libboost-chrono-dev libboost-filesystem-dev libboost-program-options-dev \
-libboost-system-dev libboost-test-dev libboost-thread-dev sudo automake git wget curl libdb4.8-dev bsdmainutils libdb4.8++-dev \
-libminiupnpc-dev libgmp3-dev unzip libzmq3-dev ufw pkg-config libevent-dev libdb5.3++>/dev/null 2>&1
-if [ "$?" -gt "0" ];
-  then
-    echo -e "${RED}Not all required packages were installed properly. Try to install them manually by running the following commands:${NC}\n"
-    echo "apt-get update"
-    echo "apt -y install software-properties-common"
-    echo "apt-add-repository -y ppa:bitcoin/bitcoin"
-    echo "apt-get update"
-    echo "apt install -y make build-essential libtool software-properties-common autoconf libssl-dev libboost-dev libboost-chrono-dev libboost-filesystem-dev \
-libboost-program-options-dev libboost-system-dev libboost-test-dev libboost-thread-dev sudo automake git curl libdb4.8-dev \
-bsdmainutils libdb4.8++-dev libminiupnpc-dev libgmp3-dev ufw fail2ban pkg-config libevent-dev"
- exit 1
-fi
-
-clear
-}
-
 
 function important_information() {
  echo
@@ -255,35 +82,38 @@ function important_information() {
 }
 
 function import_bootstrap() {
-  wget -q $COIN_BS
-  compile_error
-  COIN_ZIP=$(echo $COIN_BS | awk -F'/' '{print $NF}')
-  unzip $COIN_ZIP >/dev/null 2>&1
-  compile_error
-  cp -r ~/acedCore/blocks ~/.acedcore/blocks
-  cp -r ~/acedCore/chainstate ~/.acedcore/chainstate
-  cp -r ~/acedCore/peers.dat ~/.acedcore/peers.dat
-  rm -r ~/acedCore
-  rm $COIN_ZIP
+  if [ ! -d "$CONFIGFOLDER/blocks" ]; then
+  # Control will enter here if $DIRECTORY doesn't exist.
+    echo -e "Downloading Bootstrap..."
+    wget -q $COIN_BS
+    compile_error
+    echo -e "Importing Bootstrap..."
+    COIN_ZIP=$(echo $COIN_BS | awk -F'/' '{print $NF}')
+    unzip $COIN_ZIP
+    compile_error
+    cp -r ~/bootstrap/blocks ~/.acedcore/blocks
+    cp -r ~/bootstrap/chainstate ~/.acedcore/chainstate
+    cp -r ~/bootstrap/peers.dat ~/.acedcore/peers.dat
+    rm -r ~/bootstrap/
+    rm $COIN_ZIP
+    echo -e "Import Complete!"
+  fi
 }
 
 function setup_node() {
+  echo -e "Setting up..."
   get_ip
   create_config
   import_bootstrap
-  create_key
   update_config
-  enable_firewall
-  #install_sentinel
-  important_information
-  configure_systemd
 }
 
+function run_node() {
+  echo -e "Running node..."
+  acedd -daemon -conf=$CONFIGFOLDER/$CONFIG_FILE -datadir=$CONFIGFOLDER
+  tail -f /root/.acedcore/debug.log
+}
 
 ##### Main #####
-clear
-
-checks
-prepare_system
-compile_node
 setup_node
+run_node
